パッケージを整理。課題へ取り掛かる
