+++ src/main/java/raisetech/student/management/controller/StudentController.java
@@ -1,0 +1,32 @@
+package raisetech.student.management.controller;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+import raisetech.student.management.data.Student;
+import raisetech.student.management.data.StudentCourses;
+import raisetech.student.management.service.StudentService;
+
+@RestController
+public class StudentController {
+
+  private StudentService service;
+
+  @Autowired
+  public StudentController(StudentService service) {
+    this.service = service;
+  }
+
+  @GetMapping("/studentList")
+  public List<Student> getStudentList() {
+    return service.searchStudentList();
+
+  }
+
+  @GetMapping("/studentCoursesList")
+  public List<StudentCourses> getStudentCoursesList() {
+    return service.searchStudentCoursesList();
+  }
+
+}
