+++ src/main/java/raisetech/student/management/data/StudentCourses.java
@@ -1,5 +1,6 @@
-package raisetech.student.management;
+package raisetech.student.management.data;
 
+import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -10,7 +11,7 @@ public class StudentCourses {
   private int coursesId;
   private int studentsId;
   private String courses;
-  private String startDate;
-  private String endDate;
+  private LocalDateTime startDate;
+  private LocalDateTime endDate;
 
 }
