+++ src/main/java/raisetech/student/management/repository/StudentRepository.java
@@ -1,9 +1,12 @@
-package raisetech.student.management;
+package raisetech.student.management.repository;
 
 
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
+import raisetech.student.management.data.Student;
+import raisetech.student.management.data.StudentCourses;
+
 
 @Mapper
 public interface StudentRepository {
